Wand of Light can now Blind players
Adds a new Blind status which can be inflicted on the player with a wand of light, instead of confusion. Other sources of blindness such as darts also use the new status.

Under this effect, your accuracy for melee and ranged is greatly reduced, and worse the further away the target is. Mechanically your "to hit" is divided by a number, starting at 2 for adjacent targets, rising to 8 at max range (9 for Barachi).

Spellcasters cannot simply ignore this: spell range is also reduced, by 1 for spells with range < 5 (unless the range was only 1 anyway), by 2 for range >= 5, or by 3 for max range spells (max range is 8, so a reduction of 2 would look like only 1 on non-Barachi races). If you're a kobold other have reduced LOS you are therefore affected less.

Light is also added to the list of possible chaos effects.