As if by magic, Gadget Shops reappear
With numerous changes to items and players over time, the original
justifications for gadget shop removal no longer seem to apply, and
instead there are some compelling reasons to reintroduce them:

- Wands are no longer in such abundant supply for a player leaning
  heavily into evocations
- Misc evokables now function like unrandarts so we won't end up with
  useless duplicates (and it's no longer possible to have e.g. multiple
  boxes of beasts to use)
- Talismans are now available and randart versions exist but
  only if you're really lucky at a jewellery shop (and jewellery is
  otherwise useless for e.g. Coglins, who happen to have a strong
  evo apt)
- Orbs are awkwardly placed in an armour shop as they provide no
  defence, they are simply carried in the offhand slot
- Staves are their own object type so won't spawn in weapon shops
  and also happen to benefit from evocations skill

So this reintroduction gives us a shop that is *mainly* flavoured
around evocations, with the addition of orbs, which I feel can
convincingly be included under "gadgets".

It's a very nice shop, if you're into such niche items, and is
consequently rare, but of course available to Gozag followers for a
price.