Make "to-hit" display more accurate and consistent
A number of improvements to the % displayed while targetting and also on monster information screens.

The existing state of things was not ideal, and not made clearer by a lot of duplication of logic. In some places "to hit" was shown while in other places "to evade", meaning the player has to invert their mental arithmetic arbitrarily. This was supposedly intended to convey that some calculations did not take shield blocking into account but others did; however even this logic was not always applied and some cases of "to hit" were still pre-shield, and regardless there was no way for the average player to understand this distinction. (It also doesn't help that in code language, "to hit" is always the roll *before* accuracy penalties and shields take place). The calculations were also not done at all for thrown projectiles, and I uncovered a bug where due to a double negative, Umbra was actually *increasing* accuracy (but the % display got used the *correct* formula so nobody ever noticed).

This commit fixes all of that; shields are now taken into account in all estimations, the language is changed to consistently "to hit", throwing now also gets a % display, and Umbra is actually penalising accuracy as intended.