Refactor spell range
Simplifies and standardises various methods for determining spell range.

Added a spell_range member to actor and monster_info which will determine range for that caster specifically, rather than the old function call with an apply_bonuses parameter which was not always being used correctly - it was possible for monsters to receive the player's Vehumet range bonus in some cases (for instance I believe this was happening with Lom Lobom's Glaciate among other examples, and possibly player ghosts). Now it should be impossible for this to happen.

There was an additional parameter to remove the limit of LOS effects like darkness - this was only really used in one case, to give bouncing beams some extra range, and this itself was allowing beams to reach outside of LOS in some circumstances. So this is now handled separately and beams have an additional check to make sure they never leave the original range radius.

Monsters with a Vehumet religion are now granted the range bonus as well - I think this will only apply for player ghosts.

Moved the logic for which spells Vehumet actually increases range for into god-passive (and moved vehumet_supports_spell there too as they're both passive).

Frozen Ramparts and Noxious Bog were also special cased separately to not have a range boost so this is all centralised into the same function now. (Perhaps there should be an additional flag for no_range_modification or similar, to avoid having an ever growing list of special cases, or maybe some of the special cases are not really needed).