Wall Monsters: framework for wall dwellers
Add an HT_WALLS habitat for monsters that can occupy wall tiles; and update beams, LOS, combat, and so on to allow targetting them with all methods of dealing damage.

Effectively we pretend that the tiles are not solid as long as there is a monster there; and then treat the next immediate tile as solid regardless of its contents, to prevent firing through to the other side of the wall. Additionally they are restricted to never be more than 1 tile deep inside walls (in other words, they must always have at least 1 adjacent tile of *open* space). This prevents them hiding from the player deep in the walls, limits their ability to surprise you by jumping out of wall, and ensures they are always accessible and therefore attackable.

These type of monsters should be considered "on" the wall rather than "in" it - things swarming over the walls, or able to move through its surface, but not receiving any defensive benefit from the wall itself. Such monsters work best with some means to exploit the fact that they can surround you on every side even in a narrow corridor: working in packs, and having abilities that let them surround you and block your treat.

Wall dwellers may typically be Earth-themed due to their affinity with rock.

This commit is just the framework and concrete monsters are in futher commits.