New habitat for wall dwellers
Add an HT_WALLS habitat for monsters that can occupy wall tiles, and update beams and LOS code to allow targetting them with all normal methods of combat.

Effectively we pretend that the tiles are not solid if there is a monster there, and then treat the next immediate tile as solid regardless of its contents, to stop anything firing through to the other side of the wall.

Monsters actually using this habitat will come in following commits, but this is the framework.