Rockfish new monster
Rockfish use the new wall monster habitat to present a new post-lair/temple threat that removes the option of 1-tile corridors as a defence. In fact as they swim in rock and "flounder" on land, you are encourage to fight them in the open instead.

You encounter them as a "school of rockfish" but on being attacked they disperse into 3 separate fishes (same AC but 1/3 the health each). They will not regroup, and there is slightly more XP awarded in total for one-shotting the original form, only because it's hard to distribute the XP exactly and it'd be weird for optimum play be to *let* it disperse.