Rockfish: new wall monster
Rockfish use the new wall monster habitat to present a new early D threat that subverts the safety of 1-tile corridors.

You encounter them as a "school of rockfish" but on being attacked they disperse into 3 separate fishes (same AC but 1/3 the health each). They will not regroup, and there is slightly more XP awarded in total for one-shotting the original form, only because it's hard to distribute the XP exactly and it'd be weird for optimum play be to *let* it disperse.

This is not intended to be an earth-shatteringly original mechanic, but rather a simple way to introduce the mechanics of wall monsters early in the game and saving more extravagent mechanics for later threats.