Sigil of Binding: monster version
In preparation for a new Yaktaur threat, makes Sigil of Binding castable
by monsters.

To make this interesting as a threat and not just a potential free
escape for the player, the effect works a little differently for the
monster version. Firstly it places a greater number of the sigils, and
prefers tiles adjacent to the player rather than perferring one
adjacent and one at 2 tile distance as it is for the player. Secondly,
the duration of binding functions a bit differently: the monster
sigils bind for a random amount of time, the same as the player stepping
on their own sigil, and don't scale with spellpower (otherwise they
were trapping both player and monsters for a rather excessive amount of
time). Since they behave a little differently they are coloured
slightly differently in tiles, and are distinguished with text using x.

If the player has no other means of escape then getting trapped for a
few turns is ok; if the monster is trapped too then it'll be for a
similar amount of time but they are hasted after. If the player is able
to avoid them but the monster gets trapped, then the monster is still
only trapped for a few turns and should be able to catch up after.

Incidentally in reworking the placement logic a minor exploit was fixed:
casting when adjacent to transparent walls it was possible to place a
sigil the other side of the wall, since you.see_cell was using
LOS_DEFAULT; the new version uses actor::see_cell_no_trans which uses
LOS_NO_TRANS.