Yaktaur Scribe: new monster
A new addition to Yaktaur bands, the glyph scribe has Sigil of Binding
in its spellbook. Stat-wise they are somewhere between Captains and
plain Yaktaurs, with a penchant for robe wearing.

Sigil works well whether or not the taurs end up binding themselves,
since they can still hit the player with ranged attacks. The player
has a harder time either approaching or running away from the threat,
and the yaktaurs will be hasted after being bound and not far behind.