Silverfish: new Holy monster
The silverfish is a beloved pet of Zin, exemplifying the purity Zin
represents.

Can launch themselves past the player, doing damage and blinding on
the way past. They also are known to live in and on the very walls.

silverfish: New spellbook and silver attacks

silverfish: missed description and give it a band