Refactor spell range
Simplifies and standardises determination of spell range, fixing some bugs and removing special cases.

Added a spell_range member to actor and monster_info which will determine range for that caster specifically, rather than the old function call with an apply_bonuses parameter which was not always being used correctly - it was possible for monsters to receive the player's Vehumet range bonus in some cases (for instance I believe this was happening with Lom Lobom's Glaciate among other examples, and possibly player ghosts). Now it should be impossible for this to happen.

Beams included some confusing code to increase the beam length slightly specifically in the case of beam bounces. This necessitated extra special casing for lightning bolt to prevent it getting Vehumet's boost, required an include of spl-util.h, and meant an extra parameter to spell_range to ignore LOS effects just in this case. There was also a narrow case that lightning bolts could be made to reach *outside* LOS especially with reduced LOS from effects like robe of night.

I was able to remove this special case by having spell range always just report the *full* range and ignore LOS. Beams now automatically end if they travel further than LOS, and targetters and so on are already limited to LOS by themselves anyway.

Moved the logic for which spells Vehumet actually increases range for into god-passive (and moved vehumet_supports_spell there too as they're both passive).

Frozen Ramparts and Noxious Bog were also special cased separately to not have a range boost so this is all centralised into the same function now. (Perhaps there should be an additional flag for no_range_modification or similar, to avoid having an ever growing list of special cases, or maybe some of the special cases are not really needed).