Half Moon Moth: new aura monster
Emits an aura of magical density, making it harder for spells to travel through the ether, effectively reducing all spell ranges by half.

Can spawn in vaults and sometimes part of yaktaur bands.

Uses the old moth of suppression tile, which seems appropriate: it's a different type of suppression, but much less complicated.