Revisions to monster wand usage
The different mechanics for which monsters like or can't use wands were
fairly inscrutable and full of special casing. Monsters above HD 9 were
prevented entirely from having wands even though most wands now could
still be very relevant past beyond that point, if only the power scaled
better.

Wand power was (30+HD) which is a pretty narrow range for power really
overall and meant that at early levels wands were devastating and by
higher levels useless.

Then, to decide if a wand was "useless", monsters used a formula based
on the number of *charges players get per wand* to determine the wand
tier. This was not a great metric for this, and was quite brittle as an
implementation as it'd be very easy to change wand charges without
realising this was connected (which has happened before).

So this commit firstly adds a new set of tiers for monster wands
specifically, and this maps to which HD ranges they are useful at, with
upper and lower limits. The NO_HT_WANDS flag is removed completely:
this was *only* used for Grinder, and I just tweaked the new formula
so it produces basically the same result for any uniques at that level,
while opening up the higher tier wands to higher HDs.

Then the power scaling is given a much bigger range and scales a lot
higher with HD, and the ALWAYS_WAND flag is renamed to ARTIFICER and
gives monsters an additional wand power boost. This applies only to Ijyb
and Maurice anyway right now: the net effect of the new formula is a
slight nerf to Ijyb and a bit of a buff to Maurice. But the player is
more likely to encounter random monsters who've picked up wands later
in the dungeon and they'll know how to use them.

I moved most of the "can I use this" logic to monster::likes_wand and
removed a number of redundant checks in the process; it's now impossible
to roll digging for a monster because it's just excluded from that list
to begin with, and 0 charge wands should not exist anyway since wand
stacking changes.

It is now possible to give wands to more monsters in vaults, if
required. They could be given before, but it would just fail above HD 9.
Now you just have to give them a wand of the right tier.