Yaktaurs: Split V and D bands
Monster variation between Vaults and late D / Depths has not been helped
by a prominent and memorable group of ranged enemies being repeated in
both.

Expunging them entirely from one branch or other would be the norm for
such a diversification effort, but for Yaktaurs I feel another approach
is warranted; being indelibly (in my mind at least) a fixture of both
branches, and now that a few more options are available in terms of
support monsters to go in bands, let's try varying the band composition
enough that Yaktaur encounters *feel* wildly different in the respective
branches.

So: in D and Depths, we draw primarily from the pool of Yaktaur species
units (Captain, Cleric, Fusilier, Scribe), whilst in Vaults the 'taurs
are rather intentionally made to seem less well organised and have
instead fallen in with local denizens such as Wardens, Undying
Armouries, and Polterguardians.

These units are all designed to aid and bolster their allies in
interesting ways with plenty of potential synergies, so hopefully the
possibility of one or two of these turning up in Yaktaur packs will mix
things up sufficiently (with a few more units/abilites still in the
works to distinguish things further.)