Allow hurting with just an mid instead of actor
This eliminates a todo comment by allowing us to correctly attribute a
kill, even if it takes place in a final_effect and the subject has died
since. Results in a slight simplification of some code and more
accurate messaging.

(Part of the "player charming" series of commits, but in the end
unneccesary for that since there is no kill to attribute. Still a good
change so I left it in, but can be easily omitted.)