Rock Fish: new wall monster
Rock Fish use the new wall monster habitat to present a new threat
in early D that removes the option of 1-tile corridors as a
defence and introduces the mechanics of wall monsters to the player.

You may encounter them as a "school of rock fish" (apologies for the
additional pun) which on being defeated will disperse into 3 separate
fishes, meaning you are now attacked from multiple angles. They do not
regroup and can now be defeated individually. No XP is earned for
deafeating the school (except through banishment or pacification),
only for individual fishes.