Silverfish: new Holy wall monster
The Silverfish is a beloved pet of Zin, resplendant in their purity.

They will launch themselves past the player as a "shooting silver star",
causing silver damage and blinding you on the way; but always landing
up in melee range. They are able to swim in walls, as other Earth-based
fish are.