Balloon Yak: new Zot monster
A chaos-themed yak variant found in the realm of Zot. On death it
explodes with a wind blast and a spread of chaotic clouds, based on old
plague shambler code. During life, it sucks the player towards itself
with a great inhalation of wind.

Spawns in rather smaller packs than more mundane yaks as the clouds
would get completely over the top in large packs, or in random single
bands.

For placeholder tiles, I just recoloured the plain yak tile into red,
blue, yellow and green variants, as I couldn't decide which was the
best balloon colour (although went with red in console).

Made by a street performer wizard. Based on design by regret-index.