Revisions to monster wand usage
The different mechanics for "monster likes/dislikes or can/can't use
wand" were fairly inscrutable and full of special casing. Monsters were
arbitrarily prevented from having wands beyond hd 9 even though a good
number of wands were still capable of being relevant beyond that point
if only the wand power scaled better.

Wand power was (30+hd) which is a pretty low range for power overall and
meant that at early levels wands were devasting and at higher levels
useless.

The "is it worthless" formula was tied to how many charges players get
per wand which was also error prone if anyone changes wand charges and
doesn't realise that behaviour is connected (which has happened
before).

So this commit makes rather extensive changes in the end. Firstly a new
"tier" system for monster wands to give a min and max HD a which they're
useful. The NO_HT_WANDS flag was *only* for Grinder and is removed as
the new system produces a similar result inherently.

Then power scaling is given a bigger range and goes a lot higher with
HD; and finally the ALWAYS_WAND flag is renamed to ARTIFICER, giving
monsters an additional power boost.

This applies only to Ijyb and Maurice: the net effect is a slight nerf
to Ijyb and a slight buff to Maurice. But the player is more likely to
encounter random monsters who've picked up wands later in the dungeon
and they'll know how to use them.

I moved most of the "can I use this" logic to monster::likes_wand and
removed some redundant checks in the process; it's now impossible to
roll digging for a monster because it's a separate list, and 0 charge
wands should not have been possible to pick up anyway.

It's also possible to give wands to more monsters in vaults now, if
required.