Rock Fish: new wall monster
Rock Fish use the new wall monster habitat to present a new threat
in early D that removes the option of 1-tile corridors as a
defence and introduces the mechanics of wall monsters to the player.

You encounter them as a "school of rock fish" but on being attacked they
disperse into 3 separate fishes (same AC but 1/3 the health each). They
will not regroup, and there is slightly more XP awarded in total for
one-shotting the original form, only because it's hard to distribute
the XP exactly and it'd be weird for optimum play be to *let* it
disperse.