Comment cleanup: remove an old TODO
While it would be nice to make the referenced param const, it's nigh-on
impossible, as _monster_spell_goodness invokes tracer versions of
otherwise mutative spell functions. Changing this would require a huge
overhaul of how spells work generally, which was not really helped by
that small comment. The comment also included reference to presumably an
old naming of _monster_spell_goodness; let's just remove it.