Wall Monsters: new mechanics
Add an HT_WALLS habitat and bring back monsters that can occupy wall
tiles.

The general concept of monsters that can move in wall spaces was never
an issue but what was problematic was the execution (with Rock Worms).
Not being able to attack something with most of your tools was simply
annoying.

Beams and targetting code are all updated so wall tiles can be attacked
with all methods of combat, ranged and spells. The only exception is
cloud placement: as this would result in a lot more complexity,
it seemed reasonable to leave this as a singular combat advantage for
wall creatures.

The advantage these monsters can press is that they are able
to surround the player on all sides, even in a 1-tile corridor. Monster
implementations in following commits will be designed around mechanics
that exploit this ability further.