Wall Monsters: new mechanic
Add an HT_WALLS habitat and bring back monsters that can occupy wall
tiles.

The general *concept* of monsters that can move in wall spaces was never
the issue but what was problematic was the execution (with Rock Worms).
Most of your tools not working against something was mainly annoying.

The new iteration updates beams and targetting code so wall tiles can be
attacked with almost all methods of combat, ranged and spells. The only
exception is cloud placement: this would result in even more in-depth
changes to allow clouds to exist sometimes on walls, and it seemed
reasonable to leave this as a singular defence advantage for wall
creatures.

The advantage these monsters can press is that they are able to surround
the player on all sides, even in a 1-tile corridor. Monster
implementations in further commits will be designed around mechanics
that exploit this unique ability further.

They are unable to move on walls where any adjacent floors are tagged
no_tele_into - thus preventing them breaching walls of dangerous vaults,
or escaping from intentionally created closets (e.g. entry vaults with
transparent walls), and also markers like Tomb.

This is minorly exploitable by the player if they know about it to
restrict the monster's movement; the same is also true at the map edge.
There is a more clever solution possible for both these problems but
it's non-trivial enough that I will leave this for later improvements.