Player Charm: new effect for existing wand
Allow monsters to finally befriend the player.

Previously monsters could use the wand but would never use it on the
player; only on the player's allies to turn them against you. This made
it useless against anyone without allies, and generally a low-tier
threat especially compared to how useful it is to the player.

The new effect is designed to be a serious threat, especially from a
monster that already has a pack of their own allies, as well as
attempting to be mechanically distinct from other debuffs that tread
similar "mind control" territory (e.g. Mesmerisation or Fear).

A successful use will make all *same-species* monsters in LOS gain a
"charmer" status, with a new pink halo to indicate the perceived
alignment change.

You are tricked into thinking these monsters are your friends, and any
allies you may have will refuse to attack them. You may still attack
"charmer" monsters, but will suffer a *time* delay any time a charmer
is damaged either directly or indirectly by yourself, as you struggle
internally to come to terms with harming something you love.

They of course *don't* count as allied as far as Gods are concerned,
and it's also worth noting that even though they have a "charmer"