Balloon Yak: new Zot monster
A chaos-themed yak variant found in the realm of Zot. On death it
explodes with a wind blast and a spread of chaotic clouds, based on old
plague shambler code. During life, it sucks the player towards itself
with a great inhalation of wind.

Spawns in rather small packs or in random single bands as the clouds
would get completely over the top in large groups.

For placeholder tiles, I just recoloured the plain yak tile into red,
blue, yellow and green variants, as I couldn't decide which was the
best balloon colour (although I went with red in console as it was
available).

Made by a street performer wizard.