Floundering: show monster status
Floundering slows monster speed and causes them to fumble attacks.
Yet it's not really clear to an unspoiled player that this status even
exists, as it's not displayed anywhere; the only clue you get will be
seeing fumble messages during combat.

Adds a "(floundering)" to monster status lines viewed with 'x'.

TODO: Probably account for this in hit % as well