Make monsters not always approach on cardinals
A very old piece of code, relating back to when Crawl had circular LOS,
made monsters always move towards the cardinal approach directions in
the last few steps towards you, rather randomising the movement (as it
works further out.)

There is some debate as to whether predictable movement is better or
worse, but given that all other movement is randomised this seems
like an odd exception.

This makes things like Starburst and Rampage marginally worse; you may
need to take an extra step or two to line things up, but this doesn't
seem like a terrible consequence for abilities that are *intended* to
make you think about and manipulate positioning.

What it makes much better from a gameplay perspective is the situation
where a group of monsters in open space are all moving towards towards
you (easily observed in Ziggurats) where they all bunch into a line
and approach in almost single file, tripping over each other in the
process. This is probably a buff to monsters in these situations but
will still make area effects more satisfying for the player to use,
and will be an improvement to player allies as well.