Cobalt Lichen: new monster
Appearing in late dungeon and depths, a new wall monster threat intended
to replace Slime Creatures in D entirely. They are able to move on walls
and surround the player anywhere, and come with a number of support
monsters that apply ally buffs and player debuffs. They show up in
increasingly large packs and while unthreatening individually, getting
surrounded by a group will be no joke.

Similarly to SCs, they are a group threat with unusual mechanics that
overcome the player strategy of using corridors for protection. It would
not make sense for both to exist in the same branch(es), so this is a
great opportunity to create some more diversity between D and V. While
the choice for which should go in which could be argued either way
around, I felt that Lichen have more longevity in terms of the level
range they can be a meaningful threat in, and more variation within the
monster set due to the varying support roles. Additionally D is more
crying out for variety *in general*.

Support band members to appear at later depths will come in the next
commits; band member selection looks weird for now until it is filled
out by this.