Cobalt Lichen Sac: charming lichen support monster
Another possible member of later lichen bands, the spore sac is able to
charm you (and your allies) both with a spore melee attack, and via
spreading spore clouds released on death (reusing the plague shambler
cloud spreader again).

Spore clouds are even worse than regular charming as they can charm
you for *any* species in sight, not just that of the attacker (since
the original attacker has gone away, this was a mechanical necessity.)

An unused plant tile was repurposed for the placeholder.