Anachrobat: new Zot monster
A fast, half-bat half-humanoid acrobatic monster with high EV but low
AC compared to most Zot creatures.

They dual-wield hand cannons (with randomised brands) and use a "Bullet
Time" ability - slowing down time for itself allowing it to fire both
cannons while leaping, with the second shot bypassing whatever mechanism
you blocked the first shot with.

So if you dodge the first shot, your EV will be effectively zero for
the 2nd shot. If you block with your shield, the 2nd shot will ignore
SH. Or if the shot hits you successfully and therefore AC is applied,
the next shot will bypass AC entirely (as well as GDR).

In practise this means these monsters are worse for any build relying
mainly on a single defensive stat, and having a spread of defences
is much better as you still have a fallback when one stat is bypassed.

This ability is anachronistic not only in its origin in modern movies,
but also in that it is literally bending time. To supplement the theme,
some very out-of-place dialogue has been provided too.