Wolf Lichen Maw: lichen support monster
A ravenous lichen that has apparently grown a mouthpiece, it bites the
enemy with a composting attack. While composting, it emits a doubled
vigour aura (and also gives a regeneration enchantment to itself and
some randomly selected lichen that are below max health). The nutrients
are transferred through its network of mycelia, meaning even if one
is charmed it will still only heal lichen.

For the placeholder I raided more unused art, and combined the old mimic
mouth overlay with the mold used for other lichen.