Orb of Company: new orb ego
Adds a new effect available on orbs only: raises the summon cap of most
summoning spells by 1.

Should work well as an orb ego as it's a strong early game boost in
exchange for losing a slot but usefulness tails off later as it doesn't
apply to most of the higher level summons.

Spells that have a normal cap of 1 (for both player and monsters) are
not boosted, as these spells tend to have a good reason why they don't
allow more. Most of these spells are the more "forgecraft" type of
spells so perhaps once that exists it'll be a clearer distinction to
just boost summoning spells (with the exception of Canine Familiar).