Spirit Yak: new monster
The elegant psychopomp of the yaktaur afterlife. Has an aura that
bestows eligible allied monsters with "spirit arrows". These give ranged
weapons an additional free shot at the same target but only after a
successful hit. Being non corporeal, this shot is able to partially
bypass the target's AC.

Since the second shot only processes on a connecting hit, this is
rather more punishing for characters relying mainly on AC: characters
with good shields or dodging not only have a good chance of preventing
the additional shot, but they then still have a chance to block/dodge
the spirit arrow should they fail their first roll.

The tile is a spectralised version of the Plumootoo's "pretty yak". It
can be found in cleric bands, and can also be found wandering around
the crypt with small group of randomly picked ranged undead on their
way to the afterlife.

Still to do:
- Remove brands from spirit arrow too?
- Needs a more detailed descript of the status somewhere
- Adjust yaml stats (based on death yak too closely)
- Review monster desc & quote
- Overlay icon
- Different arrow loading message for invis mons
- Bands tailored to different places (like torpor snail)