Unify swimming logic
The check for swimming in terms of energy use was different than the
swimming check on the monster class, used for other purposes. This
meant lava swimmers (salamander etc.) were NOT considered swimming in
some places. The impact of this was minimal - just some messaging
around item drops and zombification - but a potential slipup if anyone
used the swimming check for any more significant purpose, and eliminates
an old FIXME comment.