Let dryads hang out in trees again
Now the mechanics of walls are rather more forgiving, we can
give dryads back the flavourful ability to hide in the trees.

They gain 5 SH while in the trees, rather than the old behaviour of
being practically invulnerable. This is still overall a nerf, as trees
can easily be dealt with by and number of fire-creating methods.