Cobalt Lichen: new monster
Appearing in Depths, a new wall monster threat, named for the real-life
Cobalt crust fungus.

They are able to move on walls and hence surround the player in almost
any terrain, and come with a number of support monsters that apply ally
buffs and player debuffs. They show up in increasingly large packs and
while not so threatening individually, getting surrounded by a group
will be no joke.

They are a group threat with unusual mechanics that frustrate the player
strategy of using corridors for protection.

Support band members will come in the next commits; without them these
are rather bland but they are intended to function as a multi unit
swarm.