Alistair's Pocket Shrapnel: new alchemy spell
A reversal of classic lead-to-gold alchemy (generally known as
chrysopaeia): converts actual gold coins into the baser metal silver,
in an explosive reaction which is directed towards the enemy by using
the caster's own movement. It fires in a "wide beam" of 3 parallel beams
(5 with Vehumet), moving the player forwards a step when it does so, and
can only be fired in the compass directions.

The beam uses a limited piercing mechanic called "momentum" which
pieces the first target but is stopped by the second.

It costs no MP to use as it works by unlocking the inherent magic
contained within gold; the cost of 52 gold comes from:
(level 9) x (3 beams) x (2 momentum)

This makes it a very useful spell for an alchemist's arsenal for times
when magic points run low and a little extra firepower is needed, but
expensive enough that it shouldn't ever be a main attack button.