Create Vite projects for lobby and game clients
This adds two Vite projects (a JavaScript build tool) for the lobby
and game clients, which need to be built as separate bundles.

These are just skeletons for now, files will be moved and altered in
following commits in an attempt to minimise git conflicts for anyone
who might have .js changes in a branch or fork.