Update workflows for JavaScript bundling
Add necessary steps in the Makefile and in CI to produce the .js
bundles as required when dependencies have changed, running `npm ci`
as necessary if any package.json or related files have changed.

CI pipelines have a nodejs installation added. I have not included
nodejs in deps.py, as the installation varies from system to system.
This must be documented in a further commit (once I am confident what
the documentation needs to be exactly).